Fix coutner init value.
Currently, when cluster is initialize, first is called get() operation
and if it returns any value, it's used instead of init value. However,
get() operation alwyas return a value (0 in case there's no record),
so init value is actually never used.

Use create() to initialize counter, which check if the counter exists
and if not it uses init value for initial record.