Add method for creating counter instance.
Currently, counters are created locally when allow_dirty_reads
is true or cluster-wide otherwise. In later case the call fails when the
leader is not elected (e.g. when cluster starts), which means that
turning off allow_dirty_reads cannot be actually used. When dirty reads
are allowed it also means that there's no record in log that counter was
created.

Add a method this create instance of the coutner localy and which can be
used for futher counter operations. Make getOrCreateCounter() method
cluster-wide so that when called there's always record in the log.
For backward compatibility return counter instance also from this method,
but in the future it shouldn't probably return nothing and serve only as
a method for creating counter cluster-wide.