InMemoryLog fixes
Some fixes to the InMemoryLog, specifically concurrency and resizing
mechanism.

To handle the resizing, when adding and removing entries, we rely on the
ArrayRingBuffer. Since RAFT is 1-indexed, we start the array's head at
1.

For concurrency, now we are keeping all the log's metadata behind a
read-write lock. Since we use an array to keep the log entries, we need
to keep the indexes and the actual log values behind this lock.

All the indexes handling remains the same as before. The difference is
operations must hold the proper lock and are now applied to the
ArrayRingBuffer instead of the direct array.