Handling flaky tests.
Executing the tests with taskset has almost a fixed set of tests
failing. Some others tests only rarely failed, and others I still need
to further investigate to identify the cause.

Most of the failures happen because of timeouts. We have some fixed
threads for election, handling messages, etc., which might block. I
guess this would be enough to create flaky tests when running with
taskset.

This commit is part of issue #222. Although, not all tests are fixed,
we have less room for flaky tests.