Adding replication benchmark
This is not utilizing maven submodules, that would require a lot of
work. This commits add a new subproject in the `tests` folder to run
benchmarks written with JMH.

The benchmark we added here is not "micro". It uses `RaftHandle` object
to replicate data with Raft, pretty much the way a user would. We vary
the cluster and data size in the benchmark. Manually, we can tune the
RAFT instance, changing the log implementation, fsync, etc.