Update manual generation and remove Ant
This removes all the remaning uses of Ant. For the manual generation, we
utilize the Asciidoctor Maven plugin. The generated files are output
into `target/doc` folder. We are also able to utilize the diagram
generation. This also removes all the external assets (CSS, figures).

For additional scripts execution, like the protocol tables or schema
generation, we can utilize the Maven exec plugin.