Leader step down when removing itself
After executing a membership operation where the leader removes itself,
it will step down and trigger an election round with the remaining
members.

This takes some inspiration from the Raft dissertation mechanism to
transfer the leader. After the membership operation is comitted, we know
there is a node with up-to-date log, so we can safely start an election.
This also means, after comitting a removal, we don't get below majority.

After the leader is removed, it executes the election round to elect a
new leader. This helps to reduce the unvailability, since it is not
necessary to shutdown the node to trigger the view change. We also
update the documentation regarding the membership operations.