Handle view changes during election
Created test cases for the scenarios mentioned in the GitHub issue and
the fixes.

Elected leader leaving before the voting thread has finished. This
causes a liveness issue on both ELECTION and ELECTION2. We can fix this
by stopping the voting thread before starting it again.

Scenarios the coordinator leaves before finishing the election process
and a majority still exists. ELECTION algorithm has a liveness issue in
this case, since it does not take into account changes in the view
coordinator. We handle this case by calculating if the coordinator has
changed between views, there is a majority, it is currently the
coordinator, and there is no leader elected.

Close #259.